+++ Samples/MasterPortal/Areas/EntityList/EntityListAreaRegistration.cs
@@ -3,11 +3,7 @@
   Licensed under the MIT License. See License.txt in the project root for license information.
 */
 
-using System.Web.Http;
-using System.Web.Http.OData.Routing;
-using System.Web.Http.OData.Routing.Conventions;
 using System.Web.Mvc;
-using Adxstudio.Xrm.Web.UI.EntityList.OData;
 
 namespace Site.Areas.EntityList
 {
@@ -29,24 +25,6 @@ namespace Site.Areas.EntityList
 			context.MapRoute("EntityListPackageRepositoryImage", "EntityList/PackageRepository/Images/{__portalScopeId__}/{packageImageId}", new { controller = "PackageRepository", action = "PackageImage" });
 			context.MapRoute("EntityListPackageRepositoryVersion", "EntityList/PackageRepository/Versions/{__portalScopeId__}/{packageVersionId}", new { controller = "PackageRepository", action = "PackageVersion" });
 			context.MapRoute("EntityListPackageRepositoryDiscovery", "_installer.json", new { controller = "PackageRepository", action = "GetRepositories" });
-
-			RegisterEntityListODataRoute(GlobalConfiguration.Configuration);
-		}
-
-		public void RegisterEntityListODataRoute(HttpConfiguration config)
-		{
-			config.MessageHandlers.Add(new EntityListFormatQueryMessageHandler());
-			
-			var routingConventions = ODataRoutingConventions.CreateDefault();
-			
-			routingConventions.Insert(0, new EntitySetODataRoutingConvention());
-			
-			const string routeName = "EntityListOData";
-			const string routePrefix = "_odata";
-
-			var routeConstraint = new EntityListODataPathRouteConstraint(new DefaultODataPathHandler(), routeName, routingConventions);
-			
-			config.Routes.Add(routeName, new ODataRoute(routePrefix, routeConstraint));
 		}
 	}
 }